Extract dialogs
