+++ src/pages/collators/Collators.tsx
@@ -1,203 +1,19 @@
-import { Button, Modal } from "react-daisyui";
+import { Button } from "react-daisyui";
 import { useCallback, useEffect, useMemo, useState } from "preact/hooks";
 import { h } from "preact";
 import { useNodeInfoState } from "../../NodeInfoProvider";
 import { decimalToNative, nativeToDecimal } from "../../helpers/parseNumbers";
 import { useSortBy, useTable } from "react-table";
-import { Option } from "@polkadot/types-codec";
 import { ChevronDownIcon, ChevronUpIcon } from "@heroicons/react/20/solid";
-import { PublicKey } from "../../components/PublicKey";
-import LabelledInputField from "../../components/LabelledInputField";
-import AmplitudeLogo from "../../assets/AmplitudeLogo";
-import Big from "big.js";
 import { useGlobalState } from "../../GlobalStateProvider";
 import {
   ParachainStakingCandidate,
-  ParachainStakingInflationInflationInfo,
   useStakingPallet,
 } from "../../hooks/staking/staking";
-import { getErrors, getEventBySectionAndMethod } from "../../helpers/substrate";
+import { getErrors } from "../../helpers/substrate";
 import { toast } from "react-toastify";
-
-interface DelegateToCollatorDialogProps {
-  availableBalance?: string;
-  collator?: ParachainStakingCandidate;
-  inflationInfo?: ParachainStakingInflationInflationInfo;
-  minDelegatorStake: string;
-  tokenSymbol: string;
-  visible: boolean;
-  onClose?: () => void;
-  onSubmit?: (amount: string) => void;
-}
-
-function DelegateToCollatorDialog(props: DelegateToCollatorDialogProps) {
-  const {
-    availableBalance = "0",
-    collator,
-    inflationInfo,
-    minDelegatorStake,
-    tokenSymbol,
-    visible,
-    onClose,
-    onSubmit,
-  } = props;
-
-  const [amount, setAmount] = useState<string>("");
-
-  const CollatorInfo = useMemo(
-    () =>
-      collator ? (
-        <div className="flex flex-row items-center bg-base-100 justify-between text-right px-2">
-          <div className="flex flex-row items-center">
-            <AmplitudeLogo className="w-8 h-8 mr-2" />
-            <PublicKey variant="shorter" publicKey={collator.id} />
-          </div>
-          <div>
-            <div className="text-lg">
-              APR {inflationInfo?.delegator.rewardRate.annual || "0.00%"}
-            </div>
-            <div className="text-sm text-neutral-content">
-              Min Bond {nativeToDecimal(minDelegatorStake)} {tokenSymbol}
-            </div>
-          </div>
-        </div>
-      ) : (
-        <div />
-      ),
-    [collator, inflationInfo, minDelegatorStake, tokenSymbol]
-  );
-
-  return (
-    <Modal open={visible}>
-      <Modal.Header className="font-bold">Delegate</Modal.Header>
-      <Button
-        color="ghost"
-        size="md"
-        shape="circle"
-        className="absolute right-4 top-4"
-        onClick={onClose}
-      >
-        ✕
-      </Button>
-      <Modal.Body>
-        {CollatorInfo}
-        <div className="mt-4" />
-        <LabelledInputField
-          type="number"
-          value={amount}
-          onChange={setAmount}
-          label="Amount"
-          secondaryLabel={`Available: ${nativeToDecimal(
-            availableBalance
-          ).toFixed(4)} ${tokenSymbol}`}
-          placeholder="Enter amount..."
-        />
-      </Modal.Body>
-
-      <Modal.Actions className="justify-center">
-        <Button
-          className="px-6"
-          color="primary"
-          onClick={() => onSubmit && onSubmit(amount)}
-        >
-          Delegate
-        </Button>
-      </Modal.Actions>
-    </Modal>
-  );
-}
-
-interface ConfirmDelegateDialogProps {
-  availableBalance?: string;
-  delegationAmountDecimal?: string;
-  submissionPending?: boolean;
-  tokenSymbol: string;
-  transactionFee?: Big;
-  visible: boolean;
-  onCancel?: () => void;
-  onClose?: () => void;
-  onConfirm?: () => void;
-}
-
-function ConfirmDelegateDialog(props: ConfirmDelegateDialogProps) {
-  const {
-    availableBalance = "0",
-    delegationAmountDecimal = "0",
-    tokenSymbol,
-    visible,
-    transactionFee = Big(0),
-    submissionPending = false,
-    onCancel,
-    onClose,
-    onConfirm,
-  } = props;
-
-  const balanceDecimal = nativeToDecimal(availableBalance);
-  const transactionFeeDecimal = nativeToDecimal(transactionFee.toString());
-
-  const resultingBalance = Big(balanceDecimal)
-    .minus(delegationAmountDecimal)
-    .minus(transactionFeeDecimal)
-    .toString();
-
-  return (
-    <Modal open={visible}>
-      <Modal.Header className="font-bold">Settlement Confirmation</Modal.Header>
-      <Button
-        color="ghost"
-        size="md"
-        shape="circle"
-        className="absolute right-4 top-4"
-        onClick={onClose}
-      >
-        ✕
-      </Button>
-      <Modal.Body>
-        <div className="flex flex-col items-center justify-between">
-          <div className="text-md text-neutral-content">Delegate</div>
-          <div className="text-xl mt-2">
-            {delegationAmountDecimal} {tokenSymbol}
-          </div>
-        </div>
-
-        <div className="px-4 bg-base-100 mt-8">
-          <div className="flex justify-between">
-            <span className="text-neutral-content">Available Balance</span>
-            <span>
-              {nativeToDecimal(availableBalance)} {tokenSymbol}
-            </span>
-          </div>
-          <div className="flex justify-between mt-4">
-            <span className="text-neutral-content">Fees</span>
-            <span>
-              {nativeToDecimal(transactionFee)} {tokenSymbol}
-            </span>
-          </div>
-        </div>
-        <div className="flex justify-between mt-4 px-4">
-          <span className="text-neutral-content">Resulting Balance</span>
-          <span>
-            {resultingBalance} {tokenSymbol}
-          </span>
-        </div>
-      </Modal.Body>
-
-      <Modal.Actions className="justify-center">
-        <Button className="px-6" color="ghost" onClick={onCancel}>
-          Cancel
-        </Button>
-        <Button
-          className="px-6"
-          color="primary"
-          loading={submissionPending}
-          onClick={onConfirm}
-        >
-          Confirm
-        </Button>
-      </Modal.Actions>
-    </Modal>
-  );
-}
+import DelegateToCollatorDialog from "./dialogs/DelegateToCollatorDialog";
+import ConfirmDelegateDialog from "./dialogs/ConfirmDelegateDialog";
 
 export function Collators() {
   const { api, tokenSymbol, chain } = useNodeInfoState().state;
